+++ NEWS.md
@@ -31,6 +31,8 @@ and very easy to change.
 - Add test coverage badge to README.md.
 - Check that all defined variable-codes are in data.
 - Add check that data columns are defined.
+- In `micromake()` used preserve HEADING variables, and produces px files for
+all other variables.
 
 # pxmake 0.10.0